pyproject.toml
[project]
name = "pythonanywhere-django-ninja-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "django>=5.2",
    "django-ninja>=1.4.0",
    "dotenv>=0.9.9",
    "whitenoise>=6.9.0",
    "openai-whisper>=20231117",
    "whisperx>=3.3.1",
    "torch>=2.2.0",
    "pydub>=0.25.1",
    "ffmpeg-python>=0.2.0",
    "numpy>=1.24.0",
    "huggingface_hub>=0.20.3"
]



[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=5.1.3",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-django>=4.11.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.2",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]  # E(PEP8), F(잠재적 오류), I(Import 정리-isort)
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]
mypy_path = "src"
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true

[tool.django-stubs]
django_settings_module = "_core.settings"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::django.utils.deprecation.RemovedInDjango60Warning",
    "ignore::pydantic.warnings.PydanticDeprecatedSince20",
    "ignore::DeprecationWarning"
]
